requirements.txt
streamlit
pandas
plotly
openpyxl
yfinance
google-api-python-client
google-auth
